Replace cliff by click
Cliff is not maintained and barely used (105 GitHub stars), whereas Click is
maintained and is widely used (5k+ stars on Github).

In the context of pifpaf, most of the "interesting" features of cliff are not
used (e.g. output formats), so sticking to a single parser such as click is
far enough.